backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=4.0.0a1",
    "asyncmy>=0.2.11",
    "black>=26.3.1",
    "chromadb>=1.5.5",
    "daphne>=4.2.1",
    "django>=6.0.3",
    "django-channels>=0.7.0",
    "google-genai>=1.69.0",
    "greenlet>=3.3.2",
    "lupa>=2.6",
    "mijiaapi>=3.0.5",
    "mysqlclient>=2.2.8",
    "openai>=2.30.0",
    "opentelemetry-api>=1.36.0",
    "opentelemetry-exporter-otlp-proto-http>=1.36.0",
    "opentelemetry-sdk>=1.36.0",
    "pycryptodome>=3.23.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-django>=4.12.0",
    "python-dotenv>=1.0.1",
    "redis>=7.4.0",
    "requests>=2.33.0",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.42.0",
    "wechatbot-sdk>=0.1.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-django>=4.12.0",
]
